+++ R/Deprecated.R
@@ -44,91 +44,90 @@
 #'   returnList = TRUE
 #' )
 #' }
-ee_utils_create_manifest_image_old <- function(img,
-                                           gs_uri,
-                                           assetId,
-                                           properties = NULL,
-                                           start_time = "1970-01-01",
-                                           end_time = "1970-01-01",
-                                           pyramiding_policy = 'MEAN',
-                                           returnList = FALSE,
-                                           quiet = FALSE) {
-
-  if (inherits(img, "character")) {
-    img <- read_stars(img, proxy = TRUE)
-  }
-
-  if (inherits(img, c("RasterBrick", "RasterStack", "RasterLayer"))) {
-    if (!quiet) {
-      message(
-        "Passing img from Raster* to stars-proxy, if this step take a long time\n",
-        "please use read_stars(..., proxy = TRUE) instead of raster(...)."
-      )
-      img <- stars::st_as_stars(img)
-    }
-  }
-
-  # Creating affine_transform params
-  affine_transform <- attr(img, "dimensions")
-  shear <- img %>%
-    attr("dimensions") %>%
-    attr("raster")
-  nbands <- (affine_transform$band$to - affine_transform$band$from) + 1L
-  if (length(nbands) == 0) nbands <- 1
-  band_names <- affine_transform$band$values
-  if (is.null(band_names)) band_names <- sprintf("b%s", 1:nbands)
-  name <- sprintf("projects/earthengine-legacy/assets/%s", assetId)
-
-  if (is.na(sf::st_crs(img)$wkt)) {
-    stop("x does not have a CRS defined first")
-  }
-
-  # Creating tileset
-  tilesets <- list(
-    crs = sf::st_crs(img)$wkt,
-    sources = list(
-      list(
-        uris = gs_uri,
-        affine_transform = list(
-          scale_x = affine_transform$x$delta,
-          shear_x = shear$affine[1],
-          translate_x = affine_transform$x$offset,
-          shear_y = shear$affine[2],
-          scale_y = affine_transform$y$delta,
-          translate_y = affine_transform$y$offset
-        )
-      )
-    )
-  )
-
-  # from R date to JS timestamp: time_start + time_end
-  time_start <- rdate_to_eedate(start_time, timestamp = TRUE)
-  time_end <- rdate_to_eedate(end_time, timestamp = TRUE)
-
-  # Adding bands
-  bands <- list()
-  for (index in seq_len(length(band_names))) {
-    bands[[index]] <- list(
-      id = band_names[index],
-      tileset_band_index = as.integer((index - 1))
-    )
-  }
-
-  # Putting all together
-  manifest <- list(
-    name = name,
-    tilesets = list(tilesets),
-    bands = bands,
-    pyramiding_policy = pyramiding_policy,
-    properties = properties,
-    start_time = list(seconds = time_start / 1000),
-    end_time = list(seconds = time_end / 1000)
-  )
-  if (is.null(properties)) manifest[["properties"]] <- NULL
-  if (returnList) {
-    manifest
-  } else {
-    ee_utils_create_json(json_path)
-  }
-}
-
+# ee_utils_create_manifest_image_old <- function(img,
+#                                            gs_uri,
+#                                            assetId,
+#                                            properties = NULL,
+#                                            start_time = "1970-01-01",
+#                                            end_time = "1970-01-01",
+#                                            pyramiding_policy = 'MEAN',
+#                                            returnList = FALSE,
+#                                            quiet = FALSE) {
+#
+#   if (inherits(img, "character")) {
+#     img <- stars::read_stars(img, proxy = TRUE)
+#   }
+#
+#   if (inherits(img, c("RasterBrick", "RasterStack", "RasterLayer"))) {
+#     if (!quiet) {
+#       message(
+#         "Passing img from Raster* to stars-proxy, if this step take a long time\n",
+#         "please use read_stars(..., proxy = TRUE) instead of raster(...)."
+#       )
+#       img <- stars::st_as_stars(img)
+#     }
+#   }
+#
+#   # Creating affine_transform params
+#   affine_transform <- attr(img, "dimensions")
+#   shear <- img %>%
+#     attr("dimensions") %>%
+#     attr("raster")
+#   nbands <- (affine_transform$band$to - affine_transform$band$from) + 1L
+#   if (length(nbands) == 0) nbands <- 1
+#   band_names <- affine_transform$band$values
+#   if (is.null(band_names)) band_names <- sprintf("b%s", 1:nbands)
+#   name <- sprintf("projects/earthengine-legacy/assets/%s", assetId)
+#
+#   if (is.na(sf::st_crs(img)$wkt)) {
+#     stop("x does not have a CRS defined first")
+#   }
+#
+#   # Creating tileset
+#   tilesets <- list(
+#     crs = sf::st_crs(img)$wkt,
+#     sources = list(
+#       list(
+#         uris = gs_uri,
+#         affine_transform = list(
+#           scale_x = affine_transform$x$delta,
+#           shear_x = shear$affine[1],
+#           translate_x = affine_transform$x$offset,
+#           shear_y = shear$affine[2],
+#           scale_y = affine_transform$y$delta,
+#           translate_y = affine_transform$y$offset
+#         )
+#       )
+#     )
+#   )
+#
+#   # from R date to JS timestamp: time_start + time_end
+#   time_start <- rdate_to_eedate(start_time, timestamp = TRUE)
+#   time_end <- rdate_to_eedate(end_time, timestamp = TRUE)
+#
+#   # Adding bands
+#   bands <- list()
+#   for (index in seq_len(length(band_names))) {
+#     bands[[index]] <- list(
+#       id = band_names[index],
+#       tileset_band_index = as.integer((index - 1))
+#     )
+#   }
+#
+#   # Putting all together
+#   manifest <- list(
+#     name = name,
+#     tilesets = list(tilesets),
+#     bands = bands,
+#     pyramiding_policy = pyramiding_policy,
+#     properties = properties,
+#     start_time = list(seconds = time_start / 1000),
+#     end_time = list(seconds = time_end / 1000)
+#   )
+#   if (is.null(properties)) manifest[["properties"]] <- NULL
+#   if (returnList) {
+#     manifest
+#   } else {
+#     ee_utils_create_json(json_path)
+#   }
+# }