+++ R/ee_mapViewLayersControl.R
@@ -145,16 +145,16 @@ add_basemaps <- function(map, pane = "right") {
   bgm <- c("CartoDB.Positron", "CartoDB.DarkMatter", "OpenStreetMap",
            "Esri.WorldImagery", "OpenTopoMap")
   map %>%
-    addProviderTiles(bgm[1], group = bgm[1],
-                     options = pathOptions(pane = pane)) %>%
-    addProviderTiles(bgm[2], group = bgm[2],
-                     options = pathOptions(pane = pane)) %>%
-    addProviderTiles(bgm[3], group = bgm[3],
-                     options = pathOptions(pane = pane)) %>%
-    addProviderTiles(bgm[4], group = bgm[4],
-                     options = pathOptions(pane = pane)) %>%
-    addProviderTiles(bgm[5], group = bgm[5],
-                     options = pathOptions(pane = pane)) %>%
+    leaflet::addProviderTiles(bgm[1], group = bgm[1],
+                              options = leaflet::pathOptions(pane = pane)) %>%
+    leaflet::addProviderTiles(bgm[2], group = bgm[2],
+                              options = leaflet::pathOptions(pane = pane)) %>%
+    leaflet::addProviderTiles(bgm[3], group = bgm[3],
+                              options = leaflet::pathOptions(pane = pane)) %>%
+    leaflet::addProviderTiles(bgm[4], group = bgm[4],
+                              options = leaflet::pathOptions(pane = pane)) %>%
+    leaflet::addProviderTiles(bgm[5], group = bgm[5],
+                              options = leaflet::pathOptions(pane = pane)) %>%
     leaflet::addLayersControl(
       baseGroups = bgm
     )